refactor(app): extract character, composition, and tab lifecycle handlers into hooks
- useCharacterManagement: handleOpenCharacterEditor + handleUpdateCharacter
- useCompositionState: extended with integration props to own scene/imagemap
  handle* functions (create, open, update, rename, delete)
- useTabLifecycle: handleCloseTab, processTabCloseRequest, close-left/right/
  all/others, split create/open/move/close, DnD drag-start/over/drop

App.tsx: 3864 → 3337 lines (-527). 0 lint warnings, 686 tests green.
