+++ .beads/interactions.jsonl
@@ -28,3 +28,7 @@
 {"id":"int-665fe78e","kind":"field_change","created_at":"2026-06-23T03:11:23.7711819Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-13l","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Moved handleCreateNode/handleRenameNode/handleDeleteNode/handleMoveNode/handleCut/handleCopy/handlePaste into useFileSystemManager.ts alongside the existing tree-helper primitives. App.tsx 4598->4480 lines (-118). tsc/lint/686 tests/build all green, no regressions."}}
 {"id":"int-7fc1e6ff","kind":"field_change","created_at":"2026-06-23T11:20:36.6422212Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-4oq","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Extracted getTabLabel/renderTabContent/renderTabBar into src/hooks/useTabContentRenderer.tsx via props-bag pattern matching useProjectIO/useFileSystemManager. App.tsx reduced 4480->4216 lines. tsc/lint/tests all clean, no new errors vs baseline, 686/686 tests pass."}}
 {"id":"int-d8030abb","kind":"field_change","created_at":"2026-06-23T11:56:03.2696758Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-2uu","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Folded handleAdd/Refresh/Remove ScanDirectory and CopyToProjectBulk for images+audio into useAssetManagement (params-object pattern: projectRootPath, perfRecorders, setIsScanningAssets, setHasUnsavedSettings, setFileSystemTree, addToast). App.tsx 4216->4029 lines (-187). Tests 686/686 pass, lint clean, build succeeds, tsc baseline error count unchanged (50)."}}
+{"id":"int-1589ac22","kind":"field_change","created_at":"2026-06-27T19:54:31.2716093Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-vr9","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"useDraftingArtifacts hook extracted in commit e7708e4"}}
+{"id":"int-ab6afd3e","kind":"field_change","created_at":"2026-06-27T20:02:48.9159121Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-i1s","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"handleOpenCharacterEditor/handleUpdateCharacter extracted into useCharacterManagement hook"}}
+{"id":"int-cf2974d5","kind":"field_change","created_at":"2026-06-27T20:02:49.5814437Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-1es","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Scene/imagemap handlers moved into useCompositionState; hook now accepts integration props (activeTabId, setOpenTabs, setActiveTabId, setHasUnsavedSettings)"}}
+{"id":"int-5ec55c4c","kind":"field_change","created_at":"2026-06-27T20:07:13.4185734Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-3m9","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"handleCloseTab/processTabCloseRequest/handleCreate|Close|Move|Open*Split/DnD handlers extracted into useTabLifecycle hook; App.tsx at 3337 lines (down from 3864)"}}
